+++ resources/views/index.blade.php
@@ -1,0 +1,18 @@
+@extends('layouts.app')
+
+@section('content')
+  @include('partials.page-header')
+
+  @if (!have_posts())
+    <div class="alert alert-warning">
+      {{ __('Sorry, no results were found.', 'sage') }}
+    </div>
+    {!! get_search_form(false) !!}
+  @endif
+
+  @while (have_posts()) @php(the_post())
+    @include ('partials.content-'.(get_post_type() === 'post' ?: get_post_type()))
+  @endwhile
+
+  {!! get_the_posts_navigation() !!}
+@endsection