+++ resources/views/page.blade.php
@@ -1,0 +1,8 @@
+@extends('layouts.app')
+
+@section('content')
+  @while(have_posts()) @php(the_post())
+    @include('partials.page-header')
+    @include('partials.content-page')
+  @endwhile
+@endsection