+++ resources/views/single.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.app')
+
+@section('content')
+  @while(have_posts()) @php(the_post())
+    @include('partials/content-single-'.get_post_type())
+  @endwhile
+@endsection