+++ templates/single.blade.php
@@ -1,8 +1,0 @@
-@extends('layouts.base')
-
-@section('content')
-  @while(have_posts()) @php(the_post())
-    @include('partials/content-single-'.get_post_type())
-  @endwhile
-@endsection
-